limit loop to not have unsafe downcasting uint96 in nonce step

+++ src/QuarkStateManager.sol
@@ -64,7 +64,8 @@ contract QuarkStateManager {
      * @return The next unused nonce
      */
     function nextNonce(address wallet) external view returns (uint96) {
-        for (uint256 bucket = 0; bucket < type(uint256).max;) {
+        // Exclude bucket==uint88.max out to prevent from overlfowing, as favor in using unchecked for gas optimization
+        for (uint256 bucket = 0; bucket < type(uint88).max;) {
             uint256 bucketNonces = nonces[wallet][bucket];
             uint96 bucketValue = uint96(bucket << 8);
             for (uint256 maskOffset = 0; maskOffset < 256;) {
